Normalize :title, :summary, and :body so that extra whitespace is stripped off. Also made sure that extra whitespace was stripped off from the HTML tag that contains those values.
Made it automatically focus the title attribute when the blog post is
first created. Then we they click the Continue button and go to the
detail/show page for the blog post, it will automatically focus the body
attribute.

Show "Unpublished" by/on future-dated blog posts.

Wrapped the index view and the show view in classes
(.k3_blog_blog_posts_index and .k3_blog_blog_posts_show) so that
elements inside of either of these can be styled differently if desired.